# CompProgEdu/WebApp/src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CompProgEdu API",
    "version": "v1"
  },
  "paths": {
...
    "/api/assignments/student/{courseId}": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "operationId": "GetAllStudentAssignmentsByCourseId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/AssignmentSummaryDtoListValidateableResponse"
            }
          }
        }
      }
    },
...
    "/api/courses/student/{studentId}": {
      "get": {
        "tags": [
          "Course"
        ],
        "operationId": "GetAllByStudentId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CourseSummaryDtoListValidateableResponse"
            }
          }
        }
      }
    },
...
    "/api/students": {
      "post": {
        "tags": [
          "Students"
        ],
        "operationId": "Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateStudentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoValidateableResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentGetAllResponseDtoListValidateableResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "operationId": "Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/students/{id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "operationId": "GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoValidateableResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Students"
        ],
        "operationId": "Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StudentUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/students/{id}/update-password": {
      "put": {
        "tags": [
          "Students"
        ],
        "operationId": "UpdatePassword",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateStudentPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StudentResponseDtoValidateableResponse"
            }
          }
        }
      }
    },
...
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Create",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "Delete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Update",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserGetDtoValidateableResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "fieldName": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "AssignmentSummaryDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "courseId": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentName": {
          "type": "string"
        },
        "assignmentInstructions": {
          "type": "string"
        },
        "totalPointValue": {
          "format": "int32",
          "type": "integer"
        },
        "assignmentSolution": {
          "type": "string"
        },
        "dueDate": {
          "type": "string"
        },
        "visibilityDate": {
          "type": "string"
        }
      }
    },
...
    "AssignmentSummaryDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AssignmentSummaryDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
    "UserGetDto": {
      "required": [
        "emailAddress",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "emailAddress": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
...
    "UserGetDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/UserGetDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
    "CourseSummaryDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "section": {
          "type": "string"
        },
        "studentCount": {
          "format": "int32",
          "type": "integer"
        },
        "instructorName": {
          "type": "string"
        }
      }
    },
    "CourseSummaryDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CourseSummaryDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
    "StudentResponseDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "emailAddress": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "StudentResponseDtoValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "$ref": "#/definitions/StudentResponseDto"
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
...
    "CreateStudentRequest": {
      "type": "object",
      "properties": {
        "emailAddress": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "StudentGetAllResponseDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "emailAddress": {
          "type": "string"
        },
        "grade": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        }
      }
    },
    "StudentGetAllResponseDtoListValidateableResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StudentGetAllResponseDto"
          },
          "readOnly": true
        },
        "isValidResponse": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponse"
          },
          "readOnly": true
        }
      }
    },
    "StudentUpdateDto": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "studentSchoolNumber": {
          "type": "string"
        },
        "changedPassword": {
          "type": "boolean"
        }
      }
    },
    "UpdateStudentPasswordRequest": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "newPassword": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      }
    },
...
    "CreateUserRequest": {
      "required": [
        "emailAddress",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "emailAddress": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    },
    "UserDto": {
      "required": [
        "emailAddress",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "emailAddress": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "firstName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "lastName": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "role": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer {token}"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}